+++ changelog.md
@@ -77,6 +77,8 @@
 - Fixed Creative Players defaulting to flying when entering a world
 - Removed Clair de Lune Music Disc from V2s Loot Table
 - Fixed Wing Customization Screen messing up FoV Effect Scale Setting
+- Progression Unlock Items now don't get consumed upon use while in creative
+- Progression Unlock Itemsn now use Ultracraft Titles instead of vanilla ones
 ## Edit Mode Changes
 - Rooms can now be Parented to Rooms (yay, recursion)
   - Added `/edit config recursiveRooms`, which toggles whether Rooms you pla